+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - [GUI] Don't apply previous change set to checkboxes after successful install (#2730 by: HebaruSan; reviewed: DasSkelett)
 - [Core] Compare RelationshipDescriptors with Equals instead of Same (#2735 by: HebaruSan; reviewed: DasSkelett)
 - [GUI] Auto size yes no dialog (#2729 by: HebaruSan; reviewed: DasSkelett)
+- [GUI] Fix Newly Compatible filter with repo autoupdate (#2734 by: HebaruSan; reviewed: DasSkelett)
 
 ## v1.26.0